LCD-45920 move all client code to client component
