+++ modules/util/client-extension-util-spring-boot2/src/main/java/com/liferay/client/extension/util/spring/boot2/client/ClientExtensionUtilSpringBootClientComponentScan.java
