+++ modules/util/client-extension-util-spring-boot3/src/main/java/com/liferay/client/extension/util/spring/boot3/client/ClientExtensionUtilSpringBootClientComponentScan.java
