+++ modules/util/client-extension-util-spring-boot3/src/main/java/com/liferay/client/extension/util/spring/boot3/client/LiferayOAuth2Util.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.client.extension.util.spring.boot3;
+package com.liferay.client.extension.util.spring.boot3.client;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
